ex 4.21

+++ controllers/blogs.js
@@ -47,12 +47,22 @@ blogsRouter.post("/", async (request, response) => {
 });
 
 blogsRouter.delete("/:id", async (request, response) => {
-  const deletedBlog = await Blog.findByIdAndDelete(request.params.id);
+  const decodedToken = jwt.verify(request.token, process.env.SECRET);
+  if (!decodedToken.id) {
+    return response.status(401).json({ error: "token invalid" });
+  }
 
-  if (!deletedBlog) {
+  const blog = await Blog.findById(request.params.id);
+  if (!blog) {
     return response.status(404).json({ error: "Blog not found" });
   }
 
+  // Ensure the blog's user matches the logged-in user
+  if (blog.user.toString() !== decodedToken.id) {
+    return response.status(403).json({ error: "permission denied" });
+  }
+
+  await Blog.findByIdAndDelete(request.params.id);
   response.status(204).end();
 });
